requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
pydantic>=1.9.0,<3.0.0
openai>=1.3.0
python-dotenv>=0.19.0
requests>=2.26.0
python-multipart>=0.0.5 